Basic Transfer Servlet front-end added with ajax call.
Currently getting an issue with the back-end due to the int/float changes

+++ src/main/java/edu/csudh/goTorosBank/TransferServlet.java
@@ -53,7 +53,7 @@ public class TransferServlet extends HttpServlet {
 
             int accountIDFrom = Integer.parseInt(request.getParameter("accountIDFrom"));
             int accountIDTo = Integer.parseInt(request.getParameter("accountIDTo"));
-            int amount = Integer.parseInt(request.getParameter("amount"));
+            float amount = Float.parseFloat(request.getParameter("amount"));
 
             /*Transfer the money...*/
             database.transfer(accountIDFrom, accountIDTo, amount);
